Refactoring: removed unnecessary commands (rfc templateGenerator -p service/nre -> rfc -p service/nre)
